Some file shuffling. dgemm_avx is now the blocked version, since that works better.
